Split mbcssm into independant modules
